feat: Добавлен декоратор повторного вызова фукнции в случае неудачи
